requirements.txt
google-generativeai
langchain
langchain-experimental
langchain-google-genai
langgraph
pydantic
python-dotenv
pandas
seaborn
yt-dlp
